Add files via upload
This PDF is a 53-page environmental report from the Knappen-Tippetts-Abbett (try saying that 5 times fast) Engineering Company and was commissioned by the Irvine Company in 1951, as they prepared to develop their property on the north side of the Upper and Lower Newport Bay for sale/development. It's fascinating reading. It is held, along with other related Upper/Lower Newport Bay papers and ephemera in the University of California at Irvine "Special Collections & Archives" archive:https://special.lib.uci.edu/